# CHUBB/00-Emilio-Logic/a9.location-orchestration-rest/src/main/java/com/example/location/external/Loc3x1mController.java
package com.example.location.external;

import org.springframework.http.*;
import org.springframework.web.bind.annotation.*;
import java.util.Map;

@RestController
@RequestMapping("/api/loc3x1m")
public class Loc3x1mController {
	@PostMapping("/getLocationList")
	public ResponseEntity<?> getLocationList(@RequestBody Map<String, Object> body) {
		return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED)
				.body(Map.of("faultMessageText", "LOC3X1M resolver not configured"));
	}
}