# generated/result61/src/main/java/com/locationretrievalloc3x2process/controller/http/LocStandardizeOSUSAddress2X2Controller.java
package com.locationretrievalloc3x2process.controller.http;

/**
 * Source YAML: c:/CHUBB/IA/Alberto_Recreation_Latest/TRAE-CHUBB/loc3x1-camel-route/generated/result61/src/main/resources/routes/loc-service61.yaml
 * Generated at: 2025-12-26 14:03:39
 * Usage: Implement handlers to integrate with Camel routes.
 * Steps: Validate input, route to Camel, marshal reply.
 */
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping(path = "/loc/StandardizeOSUSAddress2X2")
public class LocStandardizeOSUSAddress2X2Controller {
  private static final Logger log = LoggerFactory.getLogger(LocStandardizeOSUSAddress2X2Controller.class);

  @PostMapping(consumes = MediaType.APPLICATION_XML_VALUE, produces = MediaType.APPLICATION_XML_VALUE)
  public ResponseEntity<String> post(@RequestBody String body) {
    try {
      // TODO: implement POST handling
      return ResponseEntity.ok().contentType(MediaType.APPLICATION_XML).body("<ok/>");
    } catch (Exception e) {
      log.error("POST error: {}", e.getMessage(), e);
      return ResponseEntity.status(500).contentType(MediaType.APPLICATION_XML).body("<error>internal</error>");
    }
  }

}