# generated/result63/src/main/java/com/summary/controller/http/LocGetCountryController.java
package com.summary.controller.http;

/**
 * Source YAML: generated/result63/src/main/resources/routes/loc-service63.yaml
 * Generated at: 2025-12-29 09:23:51
 * Usage: Implement handlers to integrate with Camel routes.
 * Steps: Validate input, route to Camel, marshal reply.
 */
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping(path = "/loc/GetCountry")
public class LocGetCountryController {
  private static final Logger log = LoggerFactory.getLogger(LocGetCountryController.class);

  @PostMapping(consumes = MediaType.APPLICATION_XML_VALUE, produces = MediaType.APPLICATION_XML_VALUE)
  public ResponseEntity<String> post(@RequestBody String body) {
    try {
      // TODO: implement POST handling
      return ResponseEntity.ok().contentType(MediaType.APPLICATION_XML).body("<ok/>");
    } catch (Exception e) {
      log.error("POST error: {}", e.getMessage(), e);
      return ResponseEntity.status(500).contentType(MediaType.APPLICATION_XML).body("<error>internal</error>");
    }
  }

}